Move newMockWithWatch to kubernetes_test.go
